refactor(users): unify the domain and db entity in a single class
