+++ src/modules/users/database/user.entity.ts
@@ -1,8 +1,9 @@
-import { BaseEntity, Column, Entity, PrimaryGeneratedColumn } from 'typeorm';
+import { Column, Entity, PrimaryGeneratedColumn } from 'typeorm';
 import { UserRole } from '../util/constants/roles.ts';
+import { randomUUID } from 'crypto';
 
 @Entity('user')
-export class UserDBEntity extends BaseEntity {
+export class User {
   @PrimaryGeneratedColumn('uuid')
   id!: string;
 
@@ -14,4 +15,15 @@ export class UserDBEntity extends BaseEntity {
 
   @Column()
   password!: string;
+
+  static create(email: string, password: string, role: UserRole = UserRole.REGULAR): User {
+    const user = new User();
+
+    user.id = randomUUID();
+    user.email = email;
+    user.password = password;
+    user.role = role;
+
+    return user;
+  }
 }
